% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{ahargana}
\alias{ahargana}
\title{ahargana}
\usage{
ahargana(jd)
}
\arguments{
\item{jd}{Julian day number}
}
\value{
Ahargana
}
\description{
ahargana
}
\examples{
ahargana(2459778)
ahargana(swe_julday(2022,7,14,0,SE$GREG_CAL))
}
