% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day_duration.R
\name{day_duration}
\alias{day_duration}
\title{day_duration}
\usage{
day_duration(jd, place)
}
\arguments{
\item{jd}{Julian day number}

\item{place}{Vector containing latitude, longitude and timezone}
}
\value{
Vector containing the length of the day & in dms
}
\description{
Duration of the day for a given place and time
}
\examples{
day_duration(2459778,c(15.34, 75.13, +5.5))
day_duration(swephR::swe_julday(2022,7,14,0,swephR::SE$GREG_CAL),c(15.34, 75.13, +5.5))
}
