% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elapsed_year.R
\name{elapsed_year}
\alias{elapsed_year}
\title{elapsed_year}
\usage{
elapsed_year(jd, maasa_num)
}
\arguments{
\item{jd}{Julian Day number}

\item{maasa_num}{Number indicating the Maasa}
}
\value{
A vector containing Kali, Saka, and Vikram Samvat
}
\description{
elapsed_year
}
\examples{
elapsed_year(2459778,2)
}
