% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_dms.R
\name{from_dms}
\alias{from_dms}
\title{from_dms}
\usage{
from_dms(degs, mins, secs)
}
\arguments{
\item{degs}{Degrees}

\item{mins}{Minutes}

\item{secs}{Seconds}
}
\value{
Degrees as a decimal number
}
\description{
Convert degrees, minutes, and seconds to decimal degrees
}
\examples{
from_dms(30,15,50)
}
