% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lagna_name.R
\name{get_lagna_name}
\alias{get_lagna_name}
\title{get_lagna_name}
\usage{
get_lagna_name(jd)
}
\arguments{
\item{jd}{Julian day number}
}
\value{
Name of the lagna.
}
\description{
Get name of the Lagna for given Julian day number.
}
\examples{
get_lagna_name(2459778)
get_lagna_name(gregorian_to_jd(30,8,2022))
}
