% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_masa_name.R, R/main.R
\name{get_masa_name}
\alias{get_masa_name}
\title{get_masa_name}
\usage{
get_masa_name(jd, place)

get_masa_name(jd, place)
}
\arguments{
\item{jd}{Julian day number}

\item{place}{Vector containing latitude, longitude and timezone}
}
\value{
Name of the Masa

Name of the Masa
}
\description{
Get name of the Masa for given Julian day number and place.

Get name of the Masa for given Julian day number and place.
}
\examples{
get_masa_name(2459778,c(15.34, 75.13, +5.5))
get_masa_name(swe_julday(2022,7,14,0,SE$GREG_CAL),c(15.34, 75.13, +5.5))
get_masa_name(2459778,c(15.34, 75.13, +5.5))
get_masa_name(gregorian_to_jd(30,8,2022),c(15.34, 75.13, +5.5))
}
