% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_moonsign_name}
\alias{get_moonsign_name}
\title{get_moonsign_name}
\usage{
get_moonsign_name(jd)
}
\arguments{
\item{jd}{Julian day number}
}
\value{
Name of the Moonsign.
}
\description{
Get name of the moonsign for given Julian day number.
}
\examples{
get_moonsign_name(2459778)
get_moonsign_name(swe_julday(2022,7,14,0,SE$GREG_CAL))
}
