% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nakshatra_name.R, R/main.R
\name{get_nakshatra_name}
\alias{get_nakshatra_name}
\title{get_nakshatra_name}
\usage{
get_nakshatra_name(jd, place)

get_nakshatra_name(jd, place)
}
\arguments{
\item{jd}{Julian day number}

\item{place}{Vector containing latitude, longitude and timezone}
}
\value{
Name(s) of the Nakshatra and its ending time.

Name(s) of the Nakshatra and its ending time.
}
\description{
Get name(s) of the Nakshatra for given Julian day number and place.

Get name(s) of the Nakshatra for given Julian day number and place.
}
\examples{
get_nakshatra_name(2459778,c(15.34, 75.13, +5.5))
get_nakshatra_name(swe_julday(2022,7,14,0,SE$GREG_CAL),c(15.34, 75.13, +5.5))
get_nakshatra_name(2459778,c(15.34, 75.13, +5.5))
get_nakshatra_name(gregorian_to_jd(30,8,2022),c(15.34, 75.13, +5.5))
}
