% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_raasi_name}
\alias{get_raasi_name}
\title{get_raasi_name}
\usage{
get_raasi_name(jd)
}
\arguments{
\item{jd}{Julian day number}
}
\value{
Name of the Raasi.
}
\description{
Get name of the Raasi for given Julian day number.
}
\examples{
get_raasi_name(2459778)
get_raasi_name(swe_julday(2022,7,14,0,SE$GREG_CAL))
}
