% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rashi_name.R, R/main.R
\name{get_rashi_name}
\alias{get_rashi_name}
\title{get_rashi_name}
\usage{
get_rashi_name(jd)

get_rashi_name(jd)
}
\arguments{
\item{jd}{Julian day number}
}
\value{
Name of the Rashi.

Name of the Rashi.
}
\description{
Get name of the Rashi for given Julian day number.

Get name of the Rashi for given Julian day number.
}
\examples{
get_rashi_name(2459778)
get_rashi_name(gregorian_to_jd(30,8,2022))
get_rashi_name(2459778)
get_rashi_name(gregorian_to_jd(30,8,2022))
}
