% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ritu_name.R
\name{get_ritu_name}
\alias{get_ritu_name}
\title{get_ritu_name}
\usage{
get_ritu_name(masa_num)
}
\arguments{
\item{masa_num}{Number associated with a Masa}
}
\value{
Ritu's name
}
\description{
get_ritu_name
}
\examples{
ritu(2)
}
