% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_samvatsara_name.R
\name{get_samvatsara_name}
\alias{get_samvatsara_name}
\title{get_samvatsara_name}
\usage{
get_samvatsara_name(jd, maasa_num)
}
\arguments{
\item{jd}{Julian day number}

\item{maasa_num}{Maasa number}
}
\value{
Shaka Samvatsar
}
\description{
Name of the Shaka Samvatsar for a given Julian day number and maasa number.
}
\examples{
get_samvatsara_name(2459778,2)
}
