% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_yoga_name.R
\name{get_yoga_name}
\alias{get_yoga_name}
\title{get_yoga_name}
\usage{
get_yoga_name(jd, place)
}
\arguments{
\item{jd}{Julian day number}

\item{place}{Vector containing latitude, longitude and timezone}
}
\value{
Name(s) of the Yoga and its ending time.
}
\description{
Get name(s) of the Yoga for given Julian day number and place.
}
\examples{
get_yoga_name(2459778,c(15.34, 75.13, +5.5))
get_yoga_name(swephR::swe_julday(2022,7,14,0,swephR::SE$GREG_CAL),c(15.34, 75.13, +5.5))
}
