% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverse_lagrange.R, R/main.R
\name{inverse_lagrange}
\alias{inverse_lagrange}
\title{inverse_lagrange}
\usage{
inverse_lagrange(x, y, ya)

inverse_lagrange(x, y, ya)
}
\arguments{
\item{x}{Vector x}

\item{y}{Vector y}

\item{ya}{}
}
\value{
Value of xa

Value of xa
}
\description{
Given two vectors x and y, find the value of x = xa when y = ya, i.e., f(xa) = ya

Given two vectors x and y, find the value of x = xa when y = ya, i.e., f(xa) = ya
}
