% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/karana.R
\name{karana}
\alias{karana}
\title{karana}
\usage{
karana(jd, place)
}
\arguments{
\item{jd}{Julian day number}

\item{place}{Vector containing latitude, longitude and timezone}
}
\value{
Two karanas
}
\description{
Karana for a given place and time
}
\examples{
karana(2459778,c(15.34, 75.13, +5.5))
karana(gregorian_to_jd(17,6,2022),c(15.34, 75.13, +5.5))
}
