% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{karanas}
\alias{karanas}
\title{karanas}
\format{
An object of class \code{character} of length 60.
}
\usage{
karanas
}
\description{
Name of 60 Karanas which is when moon traverses 6° in longitude relative to the sun
}
\keyword{datasets}
