% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masa.R
\name{masa}
\alias{masa}
\title{masa}
\usage{
masa(jd, place)
}
\arguments{
\item{jd}{Julian day number}

\item{place}{Vector containing latitude, longitude and timezone}
}
\value{
Masa number and whether it is adhika masa or not
}
\description{
Masa for a given place and time
}
\examples{
masa(2459778,c(15.34, 75.13, +5.5))
masa(swephR::swe_julday(2022,7,14,0,swephR::SE$GREG_CAL),c(15.34, 75.13, +5.5))
}
