% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moon_longitude.R
\name{moon_longitude}
\alias{moon_longitude}
\title{moon_longitude}
\usage{
moon_longitude(jd)
}
\arguments{
\item{jd}{Julian day}
}
\value{
Lunar longitude for \code{jd}
}
\description{
Get Lunar longitude for a given Julian day number.
}
\examples{
moon_longitude(2459778)
moon_longitude(2459500)
}
