% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moonrise.R
\name{moonrise}
\alias{moonrise}
\title{moonrise}
\usage{
moonrise(jd, place)
}
\arguments{
\item{jd}{Julian day number}

\item{place}{Vector containing latitude, longitude and timezone}
}
\value{
Moonrise as Julian day number
}
\description{
Moonrise for a given date and place
}
\examples{
moonrise(2459778,c(15.34, 75.13, +5.5))
}
