% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moonset.R
\name{moonset}
\alias{moonset}
\title{moonset}
\usage{
moonset(jd, place)
}
\arguments{
\item{jd}{Julian day number}

\item{place}{Vector containing latitude, longitude and timezone}
}
\value{
Moonset as Julian day number
}
\description{
Moonset for a given date and place
}
\examples{
moonset(2459778,c(15.34, 75.13, +5.5))
}
