% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{moonsign}
\alias{moonsign}
\title{moonsign}
\usage{
moonsign(jd)
}
\arguments{
\item{jd}{Julian day number}
}
\value{
Moonsign as an integer
}
\description{
Moon-sign for a given Julian day number
}
\examples{
moonsign(2459778)
moonsign(swe_julday(2022,7,14,0,SE$GREG_CAL))
}
