% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nakshatra.R
\name{nakshatra}
\alias{nakshatra}
\title{nakshatra}
\usage{
nakshatra(jd, place)
}
\arguments{
\item{jd}{Julian day number}

\item{place}{Vector containing latitude, longitude and timezone}
}
\value{
Nakshatra and it's ending time
}
\description{
Nakshatra for a given place and time
}
\examples{
nakshatra(2459778,c(15.34, 75.13, +5.5))
nakshatra(gregorian_to_jd(17,6,2022),c(15.34, 75.13, +5.5))
}
