% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nakshatras}
\alias{nakshatras}
\title{nakshatras}
\format{
An object of class \code{character} of length 27.
}
\usage{
nakshatras
}
\description{
Name of the 27 Nakshatras in Vedic calendar system
}
\keyword{datasets}
