% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_moon.R
\name{new_moon}
\alias{new_moon}
\title{new_moon}
\usage{
new_moon(jd, tithi_, opt = -1)
}
\arguments{
\item{jd}{Julian day number}

\item{tithi_}{Number associated with the tithi}

\item{opt}{Option to select next new moon day(\code{opt} = 1) or previous new moon day (\code{opt} = -1), Default \code{opt} = -1 .}
}
\value{
New moon day as a Julian day number
}
\description{
Julian day representing the new moon day for a given Julian day number and tithi
}
\examples{
new_moon(2459778,2)
new_moon(2459778,tithi(2459778,c(15.34, 75.13, +5.5)))
}
