% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{raasi}
\alias{raasi}
\title{raasi}
\usage{
raasi(jd)
}
\arguments{
\item{jd}{Julian day number}
}
\value{
Raasi as an integer
}
\description{
Raasi(Sun-sign) for a given Julian day number
}
\examples{
raasi(2459778)
raasi(swe_julday(2022,7,14,0,SE$GREG_CAL))
}
