% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rashis}
\alias{rashis}
\title{rashis}
\format{
An object of class \code{character} of length 12.
}
\usage{
rashis
}
\description{
The name of 12 Rashis which represents the position of the moon on the zodiac at a given time
}
\keyword{datasets}
