% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R, R/ritu.R
\name{ritu}
\alias{ritu}
\title{ritu}
\usage{
ritu(masa_num)

ritu(masa_num)
}
\arguments{
\item{masa_num}{Number associated with a Masa}
}
\value{
Number associated with the Ritu

Number associated with the Ritu
}
\description{
ritu

ritu
}
\examples{
ritu(2)
ritu(2)
}
