% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samvatsara.R
\name{samvatsara}
\alias{samvatsara}
\title{samvatsara}
\usage{
samvatsara(jd, maasa_num)
}
\arguments{
\item{jd}{Julian day number}

\item{maasa_num}{Maasa number}
}
\value{
Number associated with the Shaka Samvatsar
}
\description{
Shaka Samvatsar for a given Julian day number and maasa number.
}
\examples{
samvatsara(2459778,2)
}
