% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{samvatsars}
\alias{samvatsars}
\title{samvatsars}
\format{
An object of class \code{character} of length 60.
}
\usage{
samvatsars
}
\description{
Name of the Year in Hindu Panchang
}
\keyword{datasets}
