% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R, R/sun_longitude.R
\name{sun_longitude}
\alias{sun_longitude}
\title{sun_longitude}
\usage{
sun_longitude(jd)

sun_longitude(jd)
}
\arguments{
\item{jd}{Julian day}
}
\value{
Solar longitude for \code{jd}

Solar longitude for \code{jd}
}
\description{
Get Solar longitude for a given Julian day number.

Get Solar longitude for a given Julian day number.
}
\examples{
sun_longitude(2459778)
sun_longitude(2459500)
sun_longitude(2459778)
sun_longitude(2459500)
}
