% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunrise.R
\name{sunrise}
\alias{sunrise}
\title{sunrise}
\usage{
sunrise(jd, place)
}
\arguments{
\item{jd}{Julian day number}

\item{place}{Vector containing latitude, longitude and timezone}
}
\value{
Sunrise as Julian day number
}
\description{
Sunrise for a given date and place
}
\examples{
sunrise(2459778,c(15.34, 75.13, +5.5))
}
