% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunset.R
\name{sunset}
\alias{sunset}
\title{sunset}
\usage{
sunset(jd, place)
}
\arguments{
\item{jd}{Julian day number}

\item{place}{Vector containing latitude, longitude and timezone}
}
\value{
Sunset as Julian day number
}
\description{
Sunset for a given date and place
}
\examples{
sunset(2459778,c(15.34, 75.13, +5.5))
}
