% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{tithi}
\alias{tithi}
\title{tithi}
\usage{
tithi(jd, place)
}
\arguments{
\item{jd}{Julian day number}

\item{place}{Vector containing latitude, longitude and timezone}
}
\value{
Tithi and it's ending time
}
\description{
Tithi for a given place and time
}
\examples{
tithi(2459778,c(15.34, 75.13, +5.5))
tithi(swe_julday(2022,6,17,0,SE$GREG_CAL),c(15.34, 75.13, +5.5))
}
