% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_dms.R
\name{to_dms}
\alias{to_dms}
\title{to_dms}
\usage{
to_dms(deg)
}
\arguments{
\item{deg}{Degrees as a decimal number}
}
\value{
A vector containing degrees, minutes and seconds
}
\description{
Convert decimal degrees to degrees, minutes, and seconds
}
\examples{
to_dms(30.263888889)
}
