% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unwrap_angles.R
\name{unwrap_angles}
\alias{unwrap_angles}
\title{unwrap_angles}
\usage{
unwrap_angles(angles)
}
\arguments{
\item{angles}{Vector containing angles}
}
\value{
angles in ascending order
}
\description{
Add 360 degs an element in the input vector if elements are not sorted in ascending order.
}
