% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaara.R
\name{vaara}
\alias{vaara}
\title{vaara}
\usage{
vaara(jd)
}
\arguments{
\item{jd}{Julian day number}
}
\value{
Vaara as an integer
}
\description{
Vaara for a given Julian day number
}
\examples{
vaara(2459778)
}
