% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yoga.R
\name{yoga}
\alias{yoga}
\title{yoga}
\usage{
yoga(jd, place)
}
\arguments{
\item{jd}{Julian day number}

\item{place}{Vector containing latitude, longitude and timezone}
}
\value{
Yoga and it's ending time
}
\description{
Yoga for a given place and time
}
\examples{
yoga(2459778,c(15.34, 75.13, +5.5))
yoga(gregorian_to_jd(17,6,2022),c(15.34, 75.13, +5.5))
}
