% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yogas}
\alias{yogas}
\title{yogas}
\format{
An object of class \code{character} of length 27.
}
\usage{
yogas
}
\description{
Name of the 27 yogas which is sum of sidereal longitudes of sun and moon in the multiples of 13 degrees 20 minutes
}
\keyword{datasets}
