% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_masa_name.R
\name{get_masa_name}
\alias{get_masa_name}
\title{get_masa_name}
\usage{
get_masa_name(jd, place)
}
\arguments{
\item{jd}{Julian day number}

\item{place}{Vector containing latitude, longitude and timezone}
}
\value{
Name of the Masa
}
\description{
Get name of the Masa for given Julian day number and place.
}
\examples{
get_masa_name(2459778,c(15.34, 75.13, +5.5))
get_masa_name(swephR::swe_julday(2022,7,14,0,swephR::SE$GREG_CAL),c(15.34, 75.13, +5.5))
}
