% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vaara_name.R
\name{get_vaara_name}
\alias{get_vaara_name}
\title{get_vaara_name}
\usage{
get_vaara_name(jd)
}
\arguments{
\item{jd}{Julian day number}
}
\value{
Name of the Vaara.
}
\description{
Get name of the Vaara for given Julian day number.
}
\examples{
get_vaara_name(2459778)
get_vaara_name(swephR::swe_julday(2022,7,14,0,swephR::SE$GREG_CAL))
}
