% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gregorian_to_jd.R
\name{gregorian_to_jd}
\alias{gregorian_to_jd}
\title{gregorian_to_jd}
\usage{
gregorian_to_jd(day, month, year)
}
\arguments{
\item{day}{Day number}

\item{month}{Month number}

\item{year}{Year number}
}
\value{
Julian day number
}
\description{
Convert Gregorian date to Julian day number at 00:00 UTC
}
\examples{
gregorian_to_jd(18,7,2022)
}
