% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd_to_gregorian.R
\name{jd_to_gregorian}
\alias{jd_to_gregorian}
\title{jd_to_gregorian}
\usage{
jd_to_gregorian(jd)
}
\arguments{
\item{jd}{Julian day number}
}
\value{
Gregorian date
}
\description{
Convert Julian day number to Gregorian date
}
\examples{
jd_to_gregorian(2459778)
}
