% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lagna.R
\name{lagna}
\alias{lagna}
\title{Lagna}
\usage{
lagna(jd)
}
\arguments{
\item{jd}{Julian day number}
}
\value{
Lagna as an integer
}
\description{
Lagna for a given Julian day number
}
\examples{
lagna(2459778)
lagna(gregorian_to_jd(30,8,2022))
}
