% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lunar_phase.R
\name{lunar_phase}
\alias{lunar_phase}
\title{lunar_phase}
\usage{
lunar_phase(jd)
}
\arguments{
\item{jd}{Julian day number}
}
\value{
Lunar phase
}
\description{
Lunar phase for a given Julian day number
}
\examples{
lunar_phase(2459778)
}
