% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rashi.R
\name{rashi}
\alias{rashi}
\title{Rashi}
\usage{
rashi(jd)
}
\arguments{
\item{jd}{Julian day number}
}
\value{
Rashi as an integer
}
\description{
Rashi for a given Julian day number
}
\examples{
rashi(2459778)
rashi(gregorian_to_jd(30,8,2022))
}
