% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R, R/tithi.R
\name{tithi}
\alias{tithi}
\title{tithi}
\usage{
tithi(jd, place)

tithi(jd, place)
}
\arguments{
\item{jd}{Julian day number}

\item{place}{Vector containing latitude, longitude and timezone}
}
\value{
Tithi and it's ending time

Tithi and it's ending time
}
\description{
Tithi for a given place and time

Tithi for a given place and time
}
\examples{
tithi(2459778,c(15.34, 75.13, +5.5))
tithi(gregorian_to_jd(17,6,2022),c(15.34, 75.13, +5.5))
tithi(2459778,c(15.34, 75.13, +5.5))
tithi(gregorian_to_jd(17,6,2022),c(15.34, 75.13, +5.5))
}
